Improve measures to prevent invalid text moves
Line moving beyond the boundaries of the buffer (or narrowed view) is
prevented, however there are still at least one an outlier case.

If the penultimate line is moved the transpose action will add an extra
newline at the end.  To stop this an additional predicate
e.g. (move-text--penultimate-line-p) will need to be made, and cleanup
will need to be done posthoc.